pyproject.toml
[project]
name = "pyenda"
version = "0.1.0"
description = "Ambiente Base-Digital"
authors = [
    {name = "Aurora",email = "gdcm10@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django (<6.0)",
    "libsass (>=0.23.0,<0.24.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
